Remove haskell package sets from shell
Because the Haskell infrastructure in nixpkgs selects one version of
each package, it can be difficult to make developer shells that work
well across multiple GHC releases. This forces us to maintain a
hand-curated set of overrides in many of our projects: a lot of work
for relatively little benefit here.

So, restrict the `devShells` to provide only GHC and cabal,
and let `cabal` fetch the appropriate dependencies for each
compiler. It is often the case that `cabal` wants to fetch alternate
versions of libraries anyway, unless you waste more time in a delicate
balancing act of selecting library versions that work for all
supported compilers.