haskell-project.nix: add zlib to shells (and option for others)
On NixOS, GHC will fail to find native libraries unless packages which
use them are plumbed into its package set. This usually manifests in a
failure to find `zlib`, because it's so ubiquitous. Add it to the
shell and provide another function argument for other native libs we
might need.